project seems to work with persistence. Need to use wordpress cli

+++ srcs/requirements/wordpress/tools/wordpress_script.sh
@@ -1,15 +1,26 @@
 #!/bin/bash
 
+echo "before if"
+
+if [ ! -d "/var/www/html/wordpress" ];
+then
+
+echo "inside if"
+
 cd /var/www/html && wget https://wordpress.org/latest.tar.gz
 
 cd /var/www/html && tar -xf latest.tar.gz
 
 rm -f /var/www/html/latest.tar.gz
 
-mv /tools/wp-config.php /var/www/html/wordpress
+cp /tools/wp-config.php /var/www/html/wordpress
+
+fi
 
 rm -f /etc/php/7.3/fpm/pool.d/www.conf
-mv /tools/www.conf /etc/php/7.3/fpm/pool.d/
+cp /tools/www.conf /etc/php/7.3/fpm/pool.d/
+
+echo "after if"
 
 chown -R www-data:www-data /var/www/html
